requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asyncio==3.4.3
atomicwrites==1.4.0
attrs==21.4.0
beautifulsoup4==4.10.0
black==21.12b0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
frozenlist==1.3.0
idna==3.3
iniconfig==1.1.1
lxml==4.7.1
multidict==5.2.0
mypy-extensions==0.4.3
numpy==1.22.1
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.1
pluggy==1.0.0
psycopg-binary==3.0.8
psycopg2==2.9.3
py==1.11.0
pyparsing==3.0.7
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
toml==0.10.2
tomli==1.2.3
typing_extensions==4.0.1
urllib3==1.26.8
yarl==1.7.2
